src/requirements.txt
pipenv
fastapi
uvicorn
psycopg2-binary
Pillow
pydantic
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
